refactor: useManage custom hook
